Refactor for multiple layouts
