+++ src/app/(admin)/AdminDashboard/page.tsx
@@ -1,0 +1,7 @@
+export default function AdminDashboard() {
+    return (
+        <>
+            <p>admining...</p>
+        </>
+    );
+}
