+++ src/app/(client)/Upload/_components/Uploader.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, DragEvent, FormEvent, createRef } from "react";
 import { fileData } from "../page";
 import { useUser } from "@clerk/nextjs";
 import { useFileContext } from "./FilesContext";
-import { MAX_FILE_SIZE, MAX_NAME_SIZE } from "~/app/api/upload/constants";
+import { MAX_FILE_SIZE, MAX_NAME_SIZE } from "~/app/(client)/api/upload/constants";
 
 export default function Uploader() {
     const { files, setFiles, submitting, setSubmitting } = useFileContext();
